+++ cli/src/main/java/uk/ac/ebi/atlas/cli/ExpressionAtlasCliCommand.java
@@ -2,12 +2,12 @@ package uk.ac.ebi.atlas.cli;
 
 import org.springframework.stereotype.Component;
 import picocli.CommandLine.Command;
-import uk.ac.ebi.atlas.cli.createExperiment.CreateExperimentCommand;
-import uk.ac.ebi.atlas.cli.experimentDesign.ExperimentDesignCommand;
+import uk.ac.ebi.atlas.cli.experiment.CreateUpdateExperimentCommand;
+import uk.ac.ebi.atlas.cli.experiment.ExperimentDesignCommand;
 
 @Command(subcommands = {
         ExperimentDesignCommand.class,
-        CreateExperimentCommand.class
+        CreateUpdateExperimentCommand.class
 })
 @Component
 public class ExpressionAtlasCliCommand {
